Refactoring the form class and form to go back towards the change set that we already get from Valkyrie.
Removes the form class and cleans up the spec tests. Corrects rubocop errors and excludes a file to the rubocop_todo yml
