+++ spec/cho/work/submissions/change_set_spec.rb
@@ -41,6 +41,7 @@ RSpec.describe Work::SubmissionChangeSet do
     before { change_set.prepopulate! }
     its(:title) { is_expected.to be_empty }
     its(:work_type) { is_expected.to be_nil }
+    its(:work_type_model) { is_expected.to be_nil }
     its(:file) { is_expected.to be_nil }
     its(:member_of_collection_ids) { is_expected.to be_empty }
   end
@@ -87,6 +88,61 @@ RSpec.describe Work::SubmissionChangeSet do
 
     it 'casts ids to Valkyrie IDs' do
       expect(change_set.member_of_collection_ids.first).to be_kind_of(Valkyrie::ID)
+    end
+  end
+
+  describe '#submit_text' do
+    its(:submit_text) { is_expected.to eq('Create Work') }
+
+    context 'save work' do
+      let(:unsaved_work) { Work::Submission.new }
+      let(:resource) { Valkyrie.config.metadata_adapter.persister.save(resource: unsaved_work) }
+
+      its(:submit_text) { is_expected.to eq('Update Work') }
+    end
+  end
+
+  describe '#form_path' do
+    its(:form_path) { is_expected.to eq('/works') }
+
+    context 'save work' do
+      let(:unsaved_work) { Work::Submission.new }
+      let(:resource) { Valkyrie.config.metadata_adapter.persister.save(resource: unsaved_work) }
+
+      its(:form_path) { is_expected.to eq("/works/#{resource.id}") }
+    end
+  end
+
+  context 'with a defined work type' do
+    let(:resource) { Work::Submission.new(work_type: work_type.id) }
+    let(:work_type) { Work::Type.find_using(label: 'Generic').first }
+
+    describe '#model' do
+      its(:model) { is_expected.to eq(resource) }
+      its(:work_type_model) { is_expected.to eq(work_type) }
+    end
+
+    describe '#fields' do
+      subject(:fields) { change_set.form_fields }
+
+      let(:work_field) { Schema::MetadataField.find_using(label: 'generic_field').first }
+
+      its(:count) { is_expected.to eq(4) }
+
+      it 'is ordered' do
+        expect(fields.map(&:label)).to eq(['title', 'subtitle', 'description', 'generic_field'])
+      end
+
+      context 'fields are reordered' do
+        before do
+          work_field.order_index = -1
+          Valkyrie.config.metadata_adapter.persister.save(resource: work_field)
+        end
+
+        it 'is ordered' do
+          expect(fields.map(&:label)).to eq(['generic_field', 'title', 'subtitle', 'description'])
+        end
+      end
     end
   end
 end
