+++ spec/views/work_submissions/new.html.erb_spec.rb
@@ -5,7 +5,7 @@ require 'rails_helper'
 RSpec.describe 'work/submissions/new', type: :view do
   let(:work_type)  { Work::Type.where(label: 'Document').first }
   let(:change_set) { Work::SubmissionChangeSet.new(Work::Submission.new(work_type: work_type.id)) }
-  let(:form) { Work::Form.new(change_set) }
+  let(:form) { change_set }
 
   before do
     assign(:work, form)
